Updated README.md as a documentation.
Considering I want the be aware of all aspects that went through
this project, I decided to use the README.md for documentation,
noting down all the structure of the website, the webpages,
and all the html containing in them in no expecific order.
This was also a good lesson for myself on how to format
the README.md file. I didn't use you could use html elements with it.